addressing the changes made
